Initial rework of repo to be a single binary
If all, or even most, ePoxy extensions are going to be similar in
nature, in terms of handling, then there should be a single binary,
running in a single container, fronted by a single GCP load balancer.
This commit is the very early, broken implementation of this.

Extensions that require separate or special handling can be created
under a cmd/ directory, and build separately.