First iteration of a working prototype of an extension "server" (#8)
* Initial rework of repo to be a single binary

If all, or even most, ePoxy extensions are going to be similar in
nature, in terms of handling, then there should be a single binary,
running in a single container, fronted by a single GCP load balancer.
This commit is the very early, broken implementation of this.

Extensions that require separate or special handling can be created
under a cmd/ directory, and build separately.

* Additional work on refactoring repo

* First iteration of working prototype of ext server

Currently, each extension directory in this repository is its own main
package, compiled separately, built into its own Docker image, run in
its own container, with its own port, own internal loadbalancer, own
systemd service unit etc. This makes is cumbersome to add new
extensions.

This commit marks the first working prototype of building an extension
"server". The main package handles requests for all extensions. The
server will be a single process, in a single container, running on a
single port.

* Gives packages and other things better names

For example, instead of package name "allocate_k8s_token", it is now
simply "token", so that thinks like "token.New()" make more sense, and
"token.Create()", etc.

* Passes a Commander to token.New()

Rather than defining a local package variable named localCommander to
implement the Commander interface, have token.New() require a Commander
to be passed to it. This way external users of the module can easily run
tests.

* Returns interface Generator from token.New() function

The user of token.New() only needs access to the methods of
k8sGenerator, so returning the interface is cleaner.

* Returns interface type Password from bmc.New()

The user of bmc.New() only uses a single method from the struct, so
set the return type of bmc.New() to the interface that the struct
implements.

* Renames various structs and functions

The old names were not intuitive. For example, it is not intuitive that
a function returning an http.Handler be named NewK8sToken. The new name
is NewTokenHandler. Same for BMC stuff.

* Uses time.Since() to calculate host uptime

time.Since() is much easier to read than time.Now().UTC().Sub(). Also
creates a constant named maxUptime instead of using 120*time.Minute
multiple times.

* Fixes a few comments

* Defines token command args in server.go, not handler.go

handler.go was not the right place for the command args. server.go
works, but we lose the hostname in the token's description, which is
probably acceptable.

This commit also introduces some name changes that are hopefully more
intuitive. All of these changes are from PR feedback.

* Shortens names of extension directories

allocate_k8s_token -> token
bmc_store_password -> bmc

Today, each of these packages only does a single thing with regard to
tokens (generates one) and bmc (stores a new password), but in the
future they could do all things related to those things, should the need
ever arise, and the shorter names are much easier and clean.

* Renames package dirs to match package names

"allocate_k8s_token" becomes just "token", and "bmc_store_password"
becomes just "bmc", and these names now match the package names they
contain.

I also renamed some interfaces and structs in the token package to
hopefully be more intuitive.

* Cleanup comments and some naming

A final review of each package revealed that in some cases comments did
not match the thing they were describing, or vice versa... and various
other small cleanup items.

* Brings test coverage to 100% for most packages

The only package that doesn't have 100% coverage is "metrics", which has
no unit tests.

* Removes redundant call to string(), already a string

* Fixes time zone issue in token unit test

The expected value in the test was using my local time. The test was
working on my local machine, but failing in Cloud Build. Adding the
--utc flag to the date command resolved this.