Adds new "delete" package
This new package will support allowing machines to request to be deleted
from the cluster. The main use case for this is MIG instances, which get
scaled in and out automatically by GCP. When a scale-in event happens
and VMs are deleted, we end up with orphaned/NotReady nodes in the
cluster.