Fixes a small bug in the "token" package
Appending to the global variable commandArgs meant that on every call to
Create(), more descriptions were being added to the array that persisted
for the life of the extension server. This small change make sure we
don't append to the global variable, but instead create a new array on
every call to Create().