Adds new "delete" package for machines to leave the cluster (#9)
* Adds new "delete" package

This new package will support allowing machines to request to be deleted
from the cluster. The main use case for this is MIG instances, which get
scaled in and out automatically by GCP. When a scale-in event happens
and VMs are deleted, we end up with orphaned/NotReady nodes in the
cluster.

* Fixes a small bug in the "token" package

Appending to the global variable commandArgs meant that on every call to
Create(), more descriptions were being added to the array that persisted
for the life of the extension server. This small change make sure we
don't append to the global variable, but instead create a new array on
every call to Create().

* Adds better logging

* Renames "delete" package to "node"

The newly renamed "node" package currently only supports a delete
function, but in the future could implement other node actions, like
labelling or tainting, etc.

* Reorganizes "node" package per PR feedback

https://github.com/m-lab/epoxy-extensions/pull/9

* Removes NodeManager interface + nodeHandler testing

Remove the unnecessary NodeManager interface, and adds tests for the
nodeManager handler.

* Removes an unnecessary commented-out import

* Context to logging + improvements from PR feedback

This commit adds some context to log message so that they can be related
to each other in the even that events happen at the same time and log
messages from multiple requests are interleaved.

Other minor improvements to the code structure based on PR feedback.

* Removes fakeCommand and Test_Newi from node_test.go

* Fixes path to 'true' command in tests

* Makes node.NewManager accept node.Command parameter

This allows us to exercise NewManager in testing without the need for a
special test case.

Also, adds contextualized logging statements to all other handlers.